Move trackerProgram epics to its own file. WIP delete stage
