+++ src/EditModel/event-program/tracker-program/program-stages/actions.js
@@ -1,4 +1,5 @@
 import eventProgramStore from '../../eventProgramStore';
+import { createActionCreator } from '../../../actions';
 import { generateUid } from 'd2/lib/uid';
 
 export const PROGRAM_STAGE_STEP_CHANGE = 'PROGRAM_STAGE_STEP_CHANGE';
@@ -12,6 +13,10 @@ export const PROGRAM_STAGE_EDIT = 'PROGRAM_STAGE_EDIT';
 export const PROGRAM_STAGE_EDIT_RESET = 'PROGRAM_STAGE_EDIT_RESET';
 export const PROGRAM_STAGE_EDIT_CANCEL = 'PROGRAM_STAGE_EDIT_CANCEL';
 export const PROGRAM_STAGE_EDIT_SAVE = 'PROGRAM_STAGE_EDIT_SAVE';
+
+export const PROGRAM_STAGE_DELETE ='PROGRAM_STAGE_DELETE';
+export const PROGRAM_STAGE_DELETE_ERROR = 'PROGRAMSTAGE_DELETE_ERROR';
+export const PROGRAM_STAGE_DELETE_SUCCESS = 'PROGRAM_STAGE_DELETE_SUCCESS';
 
 export const changeStep = stepKey => ({
     type: PROGRAM_STAGE_STEP_CHANGE,
@@ -45,3 +50,7 @@ export const cancelProgramStageEdit = () => {
         type: PROGRAM_STAGE_EDIT_CANCEL,
     };
 };
+
+export const deleteProgramStage = (stageId) => createActionCreator(PROGRAM_STAGE_DELETE)(stageId)
+export const deleteProgramStageSuccess = createActionCreator(PROGRAM_STAGE_DELETE_SUCCESS);
+export const deleteProgramStageError = createActionCreator(PROGRAM_STAGE_DELETE_ERROR);
